URL Shortner Website
This project is build on Laravel 5 framework. The main purpose of this
project is to convert the big url link to a 6 digit hash code
corresponding to the url. When user send the URL link through the form,
a 6 digit random string will  be generated and automatically connected
to the corresponding URL given by the user.